Added support for param classes
Added support for param classes in the class
example group.

This commit also changes the call to get the
description from:

self.class.meta[:example_group][:full_description]

to self.class.top_level_description.downcase

This provides better support nested descriptions.
(The previous code was concatenating the nested
descriptions which was not creating a valid
klass name for my working example)

Now, the only requirement is that the top level
description is the same as the class name.