Explicitly load functions for scope instance
Functions are loaded in Puppet as they are encountered.

If a user is just invoking functions from Ruby, then they
need to manually call the Puppet::Parser::Functions.function method
to ensure that functions are properly loaded.

This commit adds the load function call to ensure that the functions being
tested can be found by puppet.