Only create scratch DB if needed for exported resources
When writing manifests, exported resources will continue to work as
expected even without a storeconfigs DB to write to.

Previously, I was attempting to be able to check that exported resources
were exported as expected, but this is impossible without the ability to
cache the scratch DB. The scratch DB was still being created on every
run (just in case there were exported resources), but this refactors to
only created when it's truely needed (for mocking resources exported
pre-compile).

This is a speed improvement and produces no affecting changes.