Remove dependency on puppetlabs_spec_helper
Over time dependencies on the puppetlabs_spec_helper gem have appeared in the
code. This normally wouldn't be much of an issue except for the fact that
puppetlabs_spec_helper also depends on rspec-puppet and this circular
dependency wreaks havoc on package managers.

I've extracted some useful logic from puppetlabs_spec_helper (like using
Puppet::Test::TestHelper where available) and refactored the function example
group to no longer need methods provided by it.

This commit leaves the code in a bit of a messy state, but general cleanup is
soon to come.

re #56