Munge actual value into an array if expected value is an array with one item
Puppet flattens parameter values that are an array with a single value down to
just the value.  This causes significant confusion to people who expect to get
an array out if they put an array in.

So, this patch checks if the expected value is an array with only one item in
it and if so, converts the actual value into an array if it's not already one.