Fixup previous commit for readability
Without this patch the previous commit to the resource section if
coverage is at 100% is a bit difficult to read since there's a long
method chain.

This patch addresses the minor issue by breaking up the method chain and
switching to inject so the reader can see the object being built up.