Sort the untouched resources coverage list
Without this patch the list of untouched resources in the coverage
report is unsorted.  This is a problem because it's difficult to get a
sense of how many of a certain type of resource isn't covered or if a
specific resource is covered or not.

This patch addresses the problem by sorting the list of untouched
resources.