Reduce output when comparing file contents
Since file contents are diffable, the primary message line does not need
to output all the contents. In the case of short files, this is useless
duplication, in the case of longer files, it is outright harmful.

This commit fixes the message so that it doesn't output the contents of
files anymore.