Don't match compile.and_raise_error on success
Make the behavior of `compile.and_raise_error` match that of
`run.and_raise_error` and `should raise_error`: if compilation succeeds
and does not in fact raise an error, this should be considered a unit
test failure that is reported.

It would previously succeed even if no error was raised.

Fixes #276.