Allow undef parameter passing with a symbol
Now, when a symbol is used as a param value, its `inspect` method is not
called to construct the Puppet parameter value. Instead, the unquoted
string, e.g. `undef` is used in the generated resource declaration. So now
`let(:params) { { :foo => :undef } }` leads to the desired resource
declaration `class { 'test': foo => undef }`.

Before this change, it was impossible to pass `undef` as a parameter.
Using `nil`, `:undef`, or `'undef'` all failed for various reasons.

Credit to the discussion at [1] for identifying the issue and proposing
an alternate solution.

Fixes #279.

[1] <https://groups.google.com/forum/#!topic/puppet-users/6nL2eROH8is>

Adapted by Dominic Cleal from https://github.com/rodjek/rspec-puppet/pull/280
to use symbols instead of nil.