Prevent Puppet 3's _timestamp fact from invalidating cache
In Puppet 3, the Puppet::Node::Facts class adds a _timestamp fact to
the facts hash, causing constant catalogue cache invalidation between
spec examples.  This was removed in Puppet 4 via PUP-3130.

By calling #dup on the facts hash, the facts in the cache key are no
longer modified and the catalogue cache works as expected.