Respect RSpec module_path and manifest on 4x
The Puppet 4.x environment configuration hardcoded paths to a module's
fixture path; this fails when rspec-puppet is used to test entire
environments. This commit fixes this by checking to see if the relevant
settings have been set in RSpec and prefers those, and falls back to
fixture paths if they are not set.