Remove RSpec module_path config default
The RSpec `module_path` configuration option had a default value that
isn't a particularly useful default, and makes it challenging to set a
more intelligent default value. This pull request sets the `module_path`
default to nil so that a basic boolean test can be used against the
setting and a calculated default value can be used instead.