Respec RSpec config paths for code coverage
The previous mechanism for determining if resources should be added to
code coverage checks only permitted resources to be counted if they were
in a module fixture directory. This commit extracts a method for
checking if a resource should be covered, and respects the RSpec
configuration paths by allowing resources in an explicitly set module
path to be included in coverage reports.