Refactor coverage.rb for testability
The coverage class used the Singleton mixin which prevented creation of
instances of the including class. Unsurprisingly this can make testing
classes including Singleton very hard to test.

This commit refactors the coverage class to have a class level
`instance` method that conforms to the same interface as Singleton,
but also allows instances of the class to be created. It also adds a
preliminary level of test coverage so that the behavior of the class can
be verified by more than just smoke and unicorn dreams.