Use RSpec hooks for printing coverage reports
The Ruby `at_exit` hook is a bit magical, and completely unnecessary
when used with RSpec because RSpec itself provides lifecycle hooks that
are pretty much built for this sort of thing. This commit switches to
the RSpec hooks and documents using the `after(:suite)` hook for
printing a report.