Use current environment for catalogs/tests
The previous code always took an environment name, but we shouldn't let
environments switch in the middle of tests and can rely on the
configured environment found when the example is set up. This commit
adds an Adapter#current_environment method for supplying the current
environment, drops the Adapter#environment method, and drops the environment
argument for Adapter#catalog.