Configure Puppet 4.x environments at setup
When Puppet 4.x is in use, the current environments must be configured
before anything else really happens. This commit removes the logic for
setting up an environment from the Adapter4X class and moves it into the
setup_puppet method, so that when the tests are running we are
guaranteed to have valid environments.