Abstract, consistently use modulepath and manifest
In Puppet 3.x and earlier there can be an absolute modulepath and
manifest, but in Puppet 4 the module and manifest is always in the
context of the current environment. This commit adds #modulepath and #manifest
methods to Adapters::Base and overrides it in Adapters::Adapter4X with
the correct environment aware logic.

This commit also fixes a number of cases where both the modulepath and
environmentpath could have multiple colon separated directories.