Allow `adapter.manifest` to return nil
Puppet 4.x has better handling for environments that have no manifest;
instead of using file paths everywhere and magically ignoring the
manifest if it doesn't exist (like Puppet 3 does), Puppet 4 requires
that environments with no manifest be created and marked as such.
However this requires calling code to understand this and deal with it.

This commit updates the Puppet 4 adapter to check for environments with
no manifest and return nil accordingly, and update callers in Puppet 4
code paths to handle possible nil manifest values.