(ORCH-1518) Introduce app_management testing support
This patch adds support the `app_management` feature added to Puppet in 4.3.

It provides a new type `application` that can now be created much like a class
or resource. It also enables the `app_management` setting automatically when
the version of Puppet provided has this feature.

To assist with passing resource references as required by application node mappings,
this patch introduces the new `ref` helper to allow test writers to provide
resource refs that appear as raw strings (instead of being quoted like normal strings).

All this new support is documented in the README.md, and tests have been added
to provide basic coverage for this new work.

Signed-off-by: Ken Barber <ken@bob.sh>