(MODULES-3737) Allow testing internal puppet 4 functions
Some puppet 4 functions require access to the scope, which needs to be
provided through the loader.

Since the loader override needs to be active during function evaluation,
this also adds a wrapper for that func, and changes so that the func name
is displayed properly.

To reduce complexity of the rest of the code, a similar wrapper is
introduced for puppet 3. This makes the rest of the code simpler.