(MODULES-3737) fix backwards compatibility with function subject.call
Some function tests use subject.call to test for exceptions. The tests
expect call() to take an array of arguments for v3 functions. The prior
changes broke that, so this change introduces a new, uniform, way to
execute functions, deprecating the old "call()".

Originally those tests look something like this:

      expect {
        subject.call []
      }.to raise_error Puppet::ParseError, /Class foo::baz is private/

but they should be modified to look like this:

      is_expected.to run.with_params().and_raise_error(Puppet::ParseError, /Class foo::baz is private/)