Prefer spec's :type when given over the file path
Under rspec 3, the :type specified in the spec (e.g. :type => :class) is
preferred over the file location to determine the example group type.
This allows say, a class spec to be located within another directory
such as spec/functions/. Not supported under rspec 2 due to the type's
absence in the spec metadata.

Fixes #430