README: reword "Configure manifests" section
The first paragraph was hard to gasp, rewrite it to make it slightly
easier to read.
Put link to puppet reference in a list.
Update Puppet 4 reference version from 4.6 to 4.8 (latest).

Mention the conf is usually done in spec/spec_helper.rb
Add a few comments in the example configuration. Slightly refactor the
code to avoid repeating File.dirname(File.expand_path(__FILE__)).