Stub out Puppet::Util::Platform.windows?
This is a lot more complicated than it probably should have been.
Puppet::Util::Platform.windows? is used by Puppet::Util.absolute_path?
but doesn't use the Puppet windows feature to determine if we're running
on windows, instead it depends on a bit of built in Ruby arcane
knowledge (Ruby will define the constant File::ALT_SEPARATOR on Windows
and presumably other non-POSIX filesystem using OSes). So in order to
work around this, rspec-puppet will stub out all calls to this method so
that it returns true if the test case has specified a :operatingsystem
=> :windows fact, and false otherwise.

The Puppet::Util::Platform.windows? method is also used in
Puppet::Util::Windows to conditionally include a whole bunch of files
that themselves include libraries that are not available on Linux by
default (like ffi). To work around this, rspec-puppet tries to require
'puppet/util/windows' early and if it throws a LoadError, it will then
stub out any calls to Kernel.require with 'puppet/util/windows' as an
argument.

Fixes #481