Fix canonicalize_resource to only match resources with same namevar
9341e04f introduced a regression when using matchers to check that a
resource reference doesn't exist, as the matcher will find a resource of
the same type but with a different namevar/title.

The following example would fail if any `Bar` resource exists, as the
title in the `Bar[unknown]` resource reference would be ignored:

    is_expected.not_to contain_foo("foo").that_requires("Bar[unknown]")

The method passed a block to #first rather than #find in order to filter
by the namevar, but the block is unused.