Fix nil error for irretrievable `Class[main]` resource
When iterating over `@catalogue.resource_keys`, a builtin `Class[main]`
resource is returned, but isn't accessible through `#resource`, causing
a nil entry in the list of matched resources. The array of resources is
now compacted to remove inaccessible builtin resources, which fixes an
error when using `Class` resources in relationship matchers.