Support setting node parameters
This commit adds a support for setting a node's parameters. It adds a
new helper, :node_params, that is a hash with members that are passed to
the instantiated node object to become its top-scope parameters.

Currently the only ways to set top scope parameters in an example is to
specify them as facts (which isn't technically really correct and also sets
the same values in the facts hash) or set them as a :pre_condition which,
as a string, is a pain to work with when you'd like to override things
in individual examples.

Adding support for :node_params, makes it possible to set top-scope
variables for a node the same way you would when using an ENC.