Fix adapters compatibility for older Puppet 3.x
Several of the configuration parameters listed in the Adapter3X class
were not compatible with older versions of Puppet 3.x.  For puppet
modules trying to maintain compatibility with older installations, this
caused rspec to fail.  This change splits the Adapter3X class into
multiple classes for different versions, each building on the previous
to add the configuration parameters introduced in each version.