Update documentation to use Strings as Hash keys
Rubocop complains about the use of the hashrocket syntax when using Symbol
keys, but because we need to maintain Ruby 1.8.7 support I can't change the
documentation to use the 1.9 syntax.

Instead, I'll just change the documentation to use String keys, which is still
syntactically correct but avoids the Rubocop warnings for users who use the
examples verbatim in their tests.