Reset default providers for Types before compiling catalogue
Previously, if the catalogue had been converted into a RAL catalogue
(done when using the `compile.with_all_deps` matcher), the default
provider would be cached for the duration of the Ruby process and any
future `Puppet::Type` objects would have that provider set, even if the
Facter values had changed and the provider was no longer suitable.

This lead to unpredictable behaviour when switching between compiling
for different platforms and when parallelising the test execution.

The simple solution to this is to clear cached default provider for each
Puppet::Type before compiling the catalogue.