Stub out Puppet::Util::Windows::Security.supports_acl?
When testing a Windows catalogue on a *nix host, and the catalogue
contains a File resource with a mode/owner/group parameter, *and* the
catalogue has been converted into a RAL catalogue, the `windows` File
provider will call Puppet::Util::Windows::Security.supports_acl? (which
is mixed into the provider class) when validating the parameters.

This method only makes sense when applying the catalogue to an actual
host, not when unit testing the catalogue, so it can be safely stubbed
out.