puppetlabs-acl: monkey patch for non-windows o/s
This simply monkey patches the Puppet::Util::Windows::File Modules symlink? method so that on non-windows OS when running rspec-puppet on puppet code that is utilizing the puppetlabs-acl type the windows specific code it's looking for is not undefined. Any feedback is appreciated and/or changed that should be made.

/cc @rodjek