update Win32::TaskScheduler
In the lastest release of win32-taskscheduler, there was a hefty
refactor of the constants. rspec-puppet stubs these same constants and
so tests were failing because the constants had changed (yay tests).
this updates the constants to their new structure. for the most part. i
didn't write out the entire new structure for the sake of saving a few
files.