(MODULES-7854) Stub out default values for puppet6
Puppet6 modified basemodulepath so it's default value depends on the
env_windows_installdir fact. It also added a new setting,
vendormoduledir, that does the same. This commit stubs out the default
values for these settings as is done with modulepath. This behavior only
occurs when testing against Puppet6, so it's shouldn't affect older
puppet versions which don't define the vendormoduledir setting or use a
fact to determine the default basemodulepath.

See also PUP-8582.