Fix Puppet::Util::Autoload.load_file to ignore stubbing
I mistakenly added `singleton_class` when checking if the class responds to
`load_file` (to support older versions of Puppet). This is obviously not
the correct method of determining if a class method has been defined and
as such did not work as intended.

This lead to the platform stubbing still being in place when `Kernel.load`
was being called, causing the problems reported by @cdenneen (where
a provider was evaluating `Puppet.features.microsoft_windows?` at load
time).