Support directories for the manifest setting
Rspec-puppet currently doesn't support loading a directory as a manifest
like the `manifest` setting in Puppet does.

When testing environments that have a directory as a manifest you have
to resort to manually loading all the files in the directory in a
:pre_condition which leads to a clumsy experience (especially when you
need :pre_condition for other reasons).

This commit should solve this by alphabetically loading all .pp files if
manifest is a directory.