Deal with (escaped) anchors and un-anchoring
When the incoming pattern is not anchored, the filter has
to be properly loose to match (no pun intended) the dev's
intentions.

Conversely, an escaped anchor at the end needs to be
ignored. Check the test cases for how that works out.
Syntax-highlighting recommended, to see how the various
levels of back-slashes change the meaning.