Add a transitive test property
This is a minimal reproducer that's essentially A ~> B ~> C. The
that_subscribes_to matcher shows that this is equal to A ~> C but in
practice Puppet doesn't work that way. The service is not restarted if
the file is modified.

I don't know what is the expected behavior here, but it does mean that
today it's not possible to test for real world behavior.