Avoid hardcoding ipaddress6 default
In 4ef2f3643eaab642d2f24ffda6a507699d9ae0a8 this was added as a
workaround, but in 48c56051047667ff6f7f3915d7597e2da669e26d a custom
Facter implementation was added. This avoids any calls to the real
Facter and doesn't suffer from the issue where it was attempting to load
Windows gems on Linux.